authorization refactored

+++ Modules/Brand/Http/Controllers/Admin/BrandController.php
@@ -23,7 +23,7 @@ class BrandController extends Controller
 
     public function all()
     {
-//        $this->authorize('viewAny');
+        $this->authorize('viewAny');
 
         $brands = $this->brandRepository->all();
 
@@ -32,7 +32,7 @@ class BrandController extends Controller
 
     public function index()
     {
-//        $this->authorize('viewAny', Brand::class);
+        $this->authorize('viewAny', Brand::class);
 
         $brands = $this->brandRepository->jsonPaginate();
 
@@ -43,14 +43,14 @@ class BrandController extends Controller
     {
         $brand = $this->brandRepository->find($brand);
 
-//        $this->authorize('view', $brand);
+        $this->authorize('view', $brand);
 
         return new BrandResource($brand);
     }
 
     public function store(BrandCreateRequest $request)
     {
-//        $this->authorize('create',  Brand::class);
+        $this->authorize('create',  Brand::class);
 
         $brandDto = BrandDto::fromFormRequest($request);
 
@@ -63,7 +63,7 @@ class BrandController extends Controller
     {
         $brand = $this->brandRepository->find($brand);
 
-//        $this->authorize('update', $brand);
+        $this->authorize('update', $brand);
 
         $brand = $this->brandStorage->update(
             $brand, BrandDto::fromFormRequest($request)
@@ -76,7 +76,7 @@ class BrandController extends Controller
     {
         $brand = $this->brandRepository->find($brand);
 
-//        $this->authorize('delete', $brand);
+        $this->authorize('delete', $brand);
 
         $this->brandStorage->delete($brand);
